requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0

regex>=2023.12.25
python-docx>=1.1.0
PyMuPDF>=1.23.0
reportlab>=4.0.0

transformers>=4.39.0
torch>=2.2.0

pytest>=7.4.0

flask>=2.2.5
jinja2>=3.1.2
